Add starter files and setup project
Added necessary starter files to the project so
that the basic pages are set up. Added route setup
to the project so that the website will be a
single page application with client side routing.
Added basic pages and links to each. Added basic
styling to the pages for future use.